Console: change `display` to be a function
Previously, `Console.display` was expected to be an `HTMLElement` in
order to set the `innerHTML` attribute. To make the emulator more
flexible, the `display` attribute has been changed to a function that
accepts the buffer contents (defaults to `console.log`).

This allows users to output to a wider variety of targets easily.